Fix test modifying real device registry (#3703)
* Move `real_device` fixture to `conftest`

* Fix test modifying real device registry
